use long for session id

+++ src/main/java/com/richashworth/planningpoker/controller/GameController.java
@@ -31,7 +31,7 @@ public class GameController {
     }
 
     @RequestMapping("createSession")
-    public int createSession() {
+    public long createSession() {
         return sessionManager.createSession();
     }
